LayoutTests/fast/images/text-recognition/image-overlay-reversed-range-expected.txt


PASS getSelection().toString() is "foo bar"
PASS successfullyParsed is true

TEST COMPLETE
